requirements.txt
PyYAML
numpy==1.18.3
wheel
monotonic
pyobjc==6.2
webhooks
sox
flask

# google cloud API
google-cloud-speech
google-api-python-client
gcloud
pydub
pyaudio
sounddevice
google-auth

# IBM Waston
ibm-watson

# DeepAffects
deepaffects
